mm: use all of the machine's memory, not the first four gigabytes of it
A PC splits its memory around the hardware below the 4 GiB line: a few
gigabytes under it, the rest above.  The frame allocator only looked at regions
whose address fitted in 32 bits, so a machine with 8 GiB ran in 3 and never
mentioned the other 5.

Three things had to change together.  highest_address() and the free sweep take
the map as it comes.  The figures applications are given are 64-bit, because a
32-bit count of bytes cannot express 8 GiB whatever the allocator knows.  And
the self-test kept frame addresses and free-memory figures in uint32_t, which
made it fail on exactly the machines this is for.

The totals are now about memory rather than addresses.  The bitmap still spans
every address up to the top, holes included -- that is what makes it an index --
but `total` counts only the frames that are really RAM, so used and free add up
to it and a two-gigabyte hole between banks is not reported as memory in use.

  8 GiB machine:  7.9 GiB total, 13 MiB used
 12 GiB machine: 11.9 GiB total, 13 MiB used

+++ kernel/fs/ramfs.c
@@ -449,9 +449,9 @@ void ramfs_statfs(wdiskinfo_t *out)
         free_low = LOW_MEMORY_LIMIT;
 
     out->block_size   = BLOCK_SIZE;
-    out->total_bytes  = (uint32_t)(used + free_low);
-    out->free_bytes   = (uint32_t)free_low;
-    out->used_bytes   = (uint32_t)used;
+    out->total_bytes  = used + free_low;
+    out->free_bytes   = free_low;
+    out->used_bytes   = used;
     out->total_blocks = (uint32_t)((used + free_low) / BLOCK_SIZE);
     out->free_blocks  = (uint32_t)(free_low / BLOCK_SIZE);
